requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
autopep8==1.6.0
blinker==1.6.2
boto3==1.28.49
botocore==1.31.49
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.1.7
coloredlogs==15.0.1
et-xmlfile==1.1.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
frozenlist==1.4.0
humanfriendly==10.0
idna==3.3
imageio==2.31.5
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lazy_loader==0.3
llvmlite==0.41.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mpmath==1.3.0
multidict==6.0.4
mysqlclient==2.2.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
onnxruntime==1.16.1
openai==0.28.0
opencv-python-headless==4.8.1.78
openpyxl==3.0.9
packaging==23.1
pandas==1.3.4
pandas-stubs==1.2.0.35
Pillow==10.0.0
platformdirs==3.11.0
pooch==1.7.0
protobuf==4.24.4
pycodestyle==2.8.0
PyJWT==2.8.0
PyMatting==1.1.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.3
rembg==2.0.50
requests==2.26.0
s3transfer==0.6.2
scikit-image==0.22.0
scipy==1.11.3
six==1.16.0
SQLAlchemy==2.0.20
sympy==1.12
tifffile==2023.9.26
toml==0.10.2
tqdm==4.62.3
typing_extensions==4.7.1
urllib3==1.26.7
webargs==8.3.0
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2